grpc-server-async/requirements.txt
grpcio-tools==1.62.2
grpcio==1.62.2
protobuf>=4.21.6
prometheus_client==0.20.0
